split utils into bonds and formatting
