+++ src/helpers/formatting.js
@@ -1,15 +1,5 @@
-import { POOLS, ADDR_STAKING, ADX_MULTIPLIER } from "./constants"
-import { keccak256, defaultAbiCoder, bigNumberify } from "ethers/utils"
-
-export const getPool = poolId => POOLS.find(x => x.id === poolId)
-export function getBondId({ owner, amount, poolId, nonce }) {
-	return keccak256(
-		defaultAbiCoder.encode(
-			["address", "address", "uint", "bytes32", "uint"],
-			[ADDR_STAKING, owner, amount, poolId, nonce]
-		)
-	)
-}
+import { ADX_MULTIPLIER } from "./constants"
+import { bigNumberify } from "ethers/utils"
 
 export function formatADX(num) {
 	// @TODO fix this dirty hack?
